+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   not Swift 3 guidelines compliant.  
   [David Jennes](https://github.com/djbe)
   [#248](https://github.com/AliSoftware/SwiftGen/pull/248)
+* Snake cased keys with uppercase letters are correctly camel cased again.
+  [Cihat Gündüz](https://github.com/Dschee)
+  [#226](https://github.com/AliSoftware/SwiftGen/issues/226)
 
 ### New Features